requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
requests>=2.31.0
scipy>=1.11.0
Pillow>=10.0.0 